[VAS-861] feat: updated bundle list, form and details so that payment types and touchpoints manage properly the 'all' selection
