+++ src/pages/commisionalBundles/list/CommissionBundlesTableColumns.tsx
@@ -67,7 +67,7 @@ export function buildColumnDefs(t: TFunction<'translation', undefined>) {
       editable: false,
       disableColumnMenu: true,
       renderHeader: showCustomHeader,
-      renderCell: (params) => renderCell(params.row.touchpoint),
+      renderCell: (params) => renderCell(t(params.row.touchpoint)),
       sortable: false,
       flex: 4,
     },
@@ -81,7 +81,7 @@ export function buildColumnDefs(t: TFunction<'translation', undefined>) {
       editable: false,
       disableColumnMenu: true,
       renderHeader: showCustomHeader,
-      renderCell: (params) => renderCell(params.row.paymentType),
+      renderCell: (params) => renderCell(t(params.row.paymentType)),
       sortable: false,
       flex: 4,
     },
